callstack: Add a generic callstack analysis module
This module adds generic callstack, ie callstack that do not have to be
structured per pid/tid, but by any chosen elements in a trace.

This started from the current callstack analysis and views, as well as
callgraph analysis in Trace Compass 'timing' plugins and changed the
structure to support the generic callstacks. The same views from
before are still there.

It also supports user-defined callstack, such that one can define an
XML analysis that will create a callstack-like data and callgraph
can be built from it.

For the callgraph, it supports grouping results by any of the elements
of the callstack structure.

It also supports sampled data, ie data where the times are not known
but only the counts of function calls.

These plugins are meant to eventually replace Trace Compass's
callstack/graph analyses.

Change-Id: Iec229a4a4310a443218f5e3fcb8286f4b5baf603
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/98349
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>