vm: Add a new VM analysis feature
This feature adds Virtual Machine analyses to the incubator.

It started by copying the Virtual CPU Analysis module from
lttng2.kernel.core, as well as the Virtual Machine Experiment from that
same plugin.

When the incubator feature is production ready, the virtual machine
analysis and views from Trace Compass may be removed and the feature
from the incubator will replace it, unless it is desirable to eventually
have this analysis in the Trace Compass main code.

Change-Id: I5d68312b0cdd6af44a5b1f2f2482eec037b59a28
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Signed-off-by: Cédric Biancheri <cdc.biancheri@gmail.com>
Reviewed-on: https://git.eclipse.org/r/98368
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>