TraceTypes: Initial commit of TraceEvent parser
This is a trace format that can be read by Google Chrome/Chromium. It is
very interesting as each event type's behavior is defined. This makes it
able to populate views more easily.

In order to test: generate a trace in Google Chrome using the following
url: chrome://tracing

then download it and load it into the incubator.

Change-Id: Ibf8f76cc3f7060c9c0de647c103254bb6777ea92
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/98480
Reviewed-by: Hudson CI