vm: Use host ID instead of trace name
This allows to use HostThread where necessary, for instance to select a
thread to highlight.

This patch also removes the need for all traces to have a role in an
experiment before taking the events into account. This allows traces
from different physical host to be analyzed together in the same
experiment.

Change-Id: Id344e6a20605719c6fc2b7a874fa10fe39bf18ec
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/99427
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>