In and Out: Introduce the In and Out analysis
This analysis can be configured to take an input and an output then
generate a callstack, a call graph, a flame graph and statistics.

This is the first customizable analysis in tracecompass. Right click on
the analysis and select configure to set it up.

To add a new prefix use a wildcard, so *__begin will make the event
named lua_process_frame__begin display in the flame graph :
lua_process_frame.

By default it will work out of the box with "_entry" and "_exit".

It should work with experiments that have a kernel trace to set the TID
properly if the trace with the entry/exits stores the CPU it is on.

Change-Id: I56555c53e30046cf95f5b4ecd2bfd181d095de4f
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Helped-by: Marco Miller <marco.miller@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/99924
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Reviewed-by: Marco Miller <marco.miller@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/193352
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>