callstack: Remove extra callstack group
The 'callstack' element in the state system is assumed in the leaf
element. This avoids having one last group descriptor, that does not
really describe a group but an element in the state system.

Now this element is implicit, as the leaf group can also hide some other
attributes that should not be mixes with the actual callstack.

Change-Id: Ie8ced737d0c12e7ccfeb5872ba5b33adb246c78f
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/101539
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>