callstack: major refactoring of the code
Why? To use the same code path when possible for both sampling and
instrumented data

The actual callstack allows for diverse grouping of stack elements, but
preserved the same class structure as the previous callstack/callgraph
analyses in Trace Compass.

This patch proposed a restructuring of the classes, so callgraphs can
support both sampled and instrumented data, and "callstack", now known
as flame chart is just valid for instrumented.

Both callgraph and callstack use the same interfaces and base classes
for group descriptions and elements, which facilitates the callgraph
building from the instrumented data.

Change-Id: I7cf4bb2ac51024395ff158b406a59e7b3292e31f
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/102014
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>