perf: add symbol analysis
This uses the mmap[2] events to map executables and libraries to
their relative addresses and gets the symbols from the right symbols.

It is greatly inspired by the UST debug info analysis, but there is less
information available than UST, for instance the build_id of the libraries
is not present.

For now, this symbol provider works only on the machine on which the trace
was taken.

Change-Id: I1c349c569a7b8f8dd8d8a6d260729a5522452da6
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/103874
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Hudson CI