XaF: Introduce the eXtended Analysis Feature
This is the work done by Raphaël Beamonte during his PhD studies.

Most of the code will eventually move to the .core plugin (it already
uses its Activator), but there are many dependencies to the UI in many
classes.

Work to do:

* Integrate better with the rest of Trace Compass: use an XML pattern
analysis to define the model instead. It is very close to the SCXML file
used here.

* Provide better UX to display the results than the Console.

* Really separate core and ui

* unit tests

Change-Id: I2045ecf72b6d9853a5d5cad4d9458c160b296b04
Signed-off-by: Raphael Beamonte <raphael.beamonte@gmail.com>
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/105736
Reviewed-by: Hudson CI