ust: Add a userspace critical path for pthread locks
This adds a new set of plugins for extra UST analyses. The first
analysis it does is to add a handler for the critical path to handle ust
events for pthread spin lock request/acquire/unlock events.

This adds vertices and edges to the execution path and calculates the
critical path accordingly.

Change-Id: I03110a76043036cba326f1bf3a17df607ef55a91
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/111494
Reviewed-by: Hudson CI