traceevent: resolve thread ID instead of using field
This will allow various sources of trace-event events, like LTTng-UST
for JUL to resolve a different thread ID, for instance, the actual one
from the kernel. It falls back to the tid field.

It also synchronizes the getLinks method of the trace event analysis as
id caused some IOException when many threads of an opened view were
requesting the links before the analysis is built.

Change-Id: Ice079adfc553fa99d8696a5e2e86c7ffc4d0d19d
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/111918
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Hudson CI