callstack: Allow flame chart to become flat chart
If there is no CallStack attribute under the last element of the instrumented
callstack, assume the current attribute contains the status to show on one
line.

This gives access to all statistics calculated by the callstack analyses to
other types of analyses. If it's possible to get the tid of an element,
operations on threads will also be available from the flame chart view.

Change-Id: I69367fcd19d47602867cc3eed818613874246b00
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/112284
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>