vm: Introduce the virtual environment model analysis
This analysis modelises the virtual environment, ie, which machines are
involved, what's their roles and what threads are responsible for virtual
cpus.

For now, only the virtual machines are modelized, not yet the containers.

Other virtual machines analyses will use this model instead of each
rebuilding the state for themselves.

Change-Id: I5b1e71e0da804b25216e708df285e1326d0932e7
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/116378
Tested-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>