traceevent.ui: add object life span analysis
This adds support for the N-O-D events in the trace event format.

Adds a trace phase layout to all trace event analyses.

It is also (possibly) an early addopter for the Dataprovider pattern.

Icons are hand drawn.

Basic unit tests provided.

This Analysis should be considered to be uplifted out of traceevent.

Change-Id: Ida2f8f06d080566fa862822f18cba65a676ef352
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/119164
Tested-by: CI Bot
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>