callstack: Allow only one series per analysis
Allowing multiple series was a bit over-engineered. If multiple series
are possible, they can either be another level in the state system (if
they have the same structure), or XML-define the other series using the
same analysis as the first one.

With the data provider, there will be one provider per analysis/series.

Change-Id: I88634cd1bda6cea067c4dd8e4aaf7c40ad94a8ce
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/120187
Tested-by: CI Bot
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>