Add a global filters view
This adds a view where the user can enter regexes that will be applied
globally as filters on the currently opened views. The user can manage
the active global filters, delete them and save them to bring them back
later from this view.

This view is in its own plugin and can eventually be deployed as its own
feature.

Change-Id: I03b9041edcf6d54328a1f58cb142cf84a077c627
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/127218
Tested-by: CI Bot