kernel: Add failed file accesses to state system
Sometimes the file opening failed for some reason. The file access view
should show those failed attemps at access. This adds the failed return
value to a thread's file FD for the time of the open. It does not
increment the number of threads accessing the file though, as it did not
really access it.

Change-Id: If26a6607090ab4911875cfc365677ce4ae4edf72
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/133875
Tested-by: CI Bot
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>