analyses: event aspect count analyses
Event count is an on-demand analysis that counts the number of
occurrences of every event field value. It is good at finding outlyiers,
for example in an SSH log, which users connect the most often. Storing
the data locally would require orders of magnitude more space than the
original trace, so it is an on-demand analysis instead.

It is called "Event Field Count Analysis" but at the moment it uses the
aspects of a trace. This is because every aspect is a "field" in the
events table. So for a UI facing individual, this makes more sense.

TODO: handle timestamps/known aspect types

TODO: make LAMI more generalized

Change-Id: I9669f56c09d9f82b6b9585f68bf0131fbe44a5c8
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/135585
Tested-by: CI Bot
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>