filters: Edit active filter on double-click
Double-clicking a filter in the Active or Saved areas bring it back to
the edition box. If the filter was in the Active area, it is moved to
the Saved one and the filter is unapplied.

Also, since filters are per trace, when a new trace is selected, the
active filters are replaced by this trace's filters.

Change-Id: I0acd84ca08ea218b4de4cbe073b8b73888bc929e
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/147076
Tested-by: CI Bot
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>