scripting: Create data providers from all analyzes
The ScriptedAnalysis class, the one for the analyzes whose state system
is built by the script does not implement IAnalysisModule (and should
not). Both types of analyses thus don't have a common base class or
interface, so a different methods needs to be called for all types to
get data providers from them.

Change-Id: I789344a6d12be7884a3788e239ff9d2e6544dbaa
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/148892
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: CI Bot