scripting: Move the getFieldValue method to trace module
This method belongs more to the trace module than to the analysis one,
as an event is for a trace. Since they cannot share the exact same name,
the copy of the method is named getEventFieldValue.

[Deprecated] "/TraceCompass/Analysis" method getFieldValue moved to Trace

Change-Id: I287ac4d45417923eb83cf441cc9df0a2d760690b
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/151123
Tested-by: CI Bot
(cherry picked from commit 8b2cb50d88a510ca0c5b006b3bbc4317bc375212)
Reviewed-on: https://git.eclipse.org/r/151267