scripting: bug 552408 Provide scripting module help
This allows the help to appear in the Module Explorer

Also fix the javadoc, while @module work well with the EASE doclet, it
does not produce the desired result with the "normal" javadoc and causes
errors.

Also the TraceScriptingUI module in UI does not have to extend the one from
core, as it only overrides a method that is unsupported anyway in the
core.

Change-Id: Ia65f8a4b878e91cb515089ce0ca8ef101029f965
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/153668
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>