skeleton: add base ID option
This allows users to specify a base ID to replace the usual
org.eclipse.tracecompass.incubator prefix. For example, this could be
used by a company called "ACME" to create plugins for their own
features. They would then use the "com.acme" base ID.

To test this:

1. cd ./skeleton
2. ./create_new_feature.py Defaults --dir ../tracetypes --copyright Acme
3. ./create_new_feature.py Acmes --dir ../tracetypes --copyright Acme \
    --base-id com.acme

Grepping for "org.eclipse.tracecompass.incubator" proves the absence of
it, since "com.acme" was used instead in the generated "Acmes" plugin
files. Only the annotation-equipped classpath files kept it, which is
expected.

As for the "Defaults" case above, which does not use the '--base-id'
option, regenerating an existing plugin gives the same result it would
have given before, as expected.

Loading the resulting new projects in Eclipse builds successfully, same
for the maven build from clean.

Change-Id: If8528e5f0282074565856509538e046a6bf7f8af
Signed-off-by: Christophe Bedard <bedard.christophe@gmail.com>
Signed-off-by: Christophe Bedard <christophe.bedard@apex.ai>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/194973
Tested-by: Marco Miller <marco.miller@ericsson.com>
Reviewed-by: Marco Miller <marco.miller@ericsson.com>