scripting: Allow scripts to request a subset of events
Passing java objects to/from scripts can have a big overhead, especially
with the py4j engine. This adds a public ScriptEventsIterator class
which has a method that allows to filter events based on their name, so
only a subset of events can be retrieved.

With py4j, reading a big trace to get sched_switches went from 18
minutes to 43 seconds.

[added] Allow scripts to request only a subset of events

Change-Id: Ia2c1ab27854dda24be9c4ac190710bd8ddf1e277
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/154910
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>