scripting: Allow scripts to define more than one data provider
Previously, the data provider name was the analysis name, which meant
only one data provider per analysis was possible. We add a dpName
parameter to all the data provider creation methods to specify a name
for the data provider, which allows analyses to define as many data
providers as they want.

[Added] Possibility to specify the data provider name in creation methods, to allow analyses to have more than one data provider

Change-Id: I330af099782709ab633e212c11ec9987727d5183
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/158613
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>