trace-server: Update experiment and trace POST status and effect
When posting an experiment or trace identical to one on the server, return an
OK status with the experiment or trace.

When posting an experiment with a name that exists, but with a different
trace set, return a CONFLICT status and return the original experiment
on server.

For the trace, it's the path that is the key to the trace, so a trace
with identical name but different path will create a new trace without
problem.

Add unit tests for these situations.

Change-Id: I322270edb4465e6731892d052d505991b70f5f4f
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/163272
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>