trace-server: Create experiment in workspace and detect conflicts
Create an experiment resource in the workspace's Tracing project. Add
its trace resources and create its supplementary folder. Set the
experiment type and trace types.

Detect a conflict when an experiment with the same name exists with a
different set of traces. Detect the conflict even if the experiment is
not opened.

Create the trace resource using the given trace name, not the last
segment of the given path. Detect a conflict when a trace with the same
name exists in the same parent folder but with a different target
location (path) or a different trace type.

Posting a trace only creates the trace in the workspace and detects its
trace type. A trace instance is not created. Trace instances are created
only when the experiment is posted.

Delete the experiment's resources and supplementary folder when the
experiment is deleted.

Delete the traces resources and supplementary folder when a
trace is deleted only if the trace is not used in any remaining
experiment. Otherwise reject the trace deletion.

Assign a fixed UUID based on the trace path and name in
TraceManagerService.
Assign a fixed UUID based on the experiment name in
ExperimentManagerService.

Fix the endpoints in DataProviderService to use expUUID instead of uuid,
according to TSP specification. Get the experiment instance from the
ExperimentManagerService.

Create new classes Trace and Experiment for the JSon model. This is
required in order to use the trace server assigned trace UUID and
experiment UUID in TSP response, instead of the instance's getUUID().

Update unit tests to use experiments when appropriate according to TSP
specification.

Change-Id: Ic749939e4e26a3e9f3ac8d625fa0d173b5334801
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/170789
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>