rocm: Adds the support for ROCm trace type using CTF
ROCm is a collection of libraries that enable GPU computing with AMD
hardware. This patch supports traces generated using ROC-tracer and
ROC-profiler in a CTF format. This patch provides a new trace type
for these traces, and a call stack analysis for the different APIs
used in the ROCm environment.

Signed-off-by: Arnaud Fiorini <fiorini.arnaud@gmail.com>
Change-Id: Ie0ca1916ec6559ee67605dff0b75bf57e71f4818
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/175012
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>