trace-server: Return all created experiments not only opened experiments
Restore the list of created traces and experiments at initialization.

Return all created experiments in GET /experiments.

Instantiate the experiment when any request is received for an already
created but closed experiment, e.g. following a restart.

Allow creation of Experiment and Trace models from their workspace
resource, in which case nbEvents, start and end are set to 0L and
indexing status is set to "CLOSED".

Change trace UUID to be based on its resource linked location and name
so that it can be computed from the resource of a closed trace.

Change-Id: I1ceec42ebf2e7e20b8a90f29bdbf1a3f49530057
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/176364
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>