otf2: Add support for OTF2 traces converted in CTF
This patch supports OTF2 traces generated with scorep
and converted in CTF. A new tracetype, a callstack and a communicators
analysis are provided. Enter/exit and MPI
communications events are supported.
Links to the converter and to test traces (to DORSAL lab git)
are provided in the documentation of the plugin

Signed-off-by: yoann-heitz <yoann.heitz@polymtl.ca>
Change-Id: I3b26e506c9a0beb5d8f5bb798545710cba65d9d1
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/179562
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>