trace-server: simplify server classes extension
Before this commit, an user interested in creating
an own custom trace-server, had to create an
own Application class and override multiple
methods of that class to provide an own WebApplication.

After this commit, the user can easily extend the
Application class, by providing a WebApplication
as constructor.

The WebApplication class is also enriched with
a new method that can be overridden by subclasses,
to simply specify custom resources/endpoints and
serializers.

Change-Id: I0e89dada9c0bd78aeb43de9a89106516334519c5
Signed-off-by: Francesco Robino <francesco.robino@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/182128
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>