server.test: query known entry and state for timegraph tooltip test
Before it used the last row return from Iterators.getLast() of a Set of
rows and hence, the order was not determined. Moreover, the data
provider implementation changed in Trace Compass and now it returns
a value depending on the row and not an empty map.

This patch makes sure that always a known row is used which leads
to a constant tooltip for query time in the test.

Moreover, the test also queries the tooltip for arrows and annotations.
The old data provider implementation returned an empty tooltip map. The
new implementation, returns the tooltip of the state for the row and
time (Bug 575843). So, this test requires the patch for this test
to succeed:
https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/185062

Change-Id: Ieab82184cd4d58c4343765569dfc43aff18e0820
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/185064
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>