releng: Apply 2020-06 patch to init 2020-06 branch
Init this new branch this way now, as swagger-core could only be added
to the master target. Before this change, the 2020-06 target failed to
build because of that addition. There was no way found to build 2020-06
along while adding swagger-core to it. Hence this branching off of
master right before the swagger addition.

This change matches the result of [1] below, thus also removes the no
longer needed patch file, in this branch. This commit also makes 2020-06
the default target-platform, in main pom.

The CI job [2] (below) currently expects this branch to be built using
[3] or similar. Below, [3]'s properties are set by means of changing the
main pom file accordingly as well.

[1] git apply patches/0001-releng-Fix-build-with-tracecompass-incubator-2020-06.patch
[2] https://ci.eclipse.org/tracecompass/job/tracecompass-incubator-nightly-2020-06
[3] mvn clean install -Djdk.version=1.8 -Djdk.release=8 -Dtycho-use-project-settings=false

Change-Id: I5db477e83bbd23722337340fe98939db20c19f33
Signed-off-by: Marco Miller <marco.miller@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/185912
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>