server: Init OpenAPIDefinition swagger annotations
Base these annotations on below [1]'s manually documented content. This
change is part of solving [2] among other changes and more to come. It
is based on [3]. Reference [4]'s README change explains how to test it.

Add these constant yet related strings using EndpointConstants.

Formatting is the default one coming with this Eclipse project.

Comparing [1]'s contents with this change, only typos found in [1] were
fixed herein. The original content previously documented through [1] is
kept in this change, for this initial step that is. There are more such
annotations to come, to keep aligning this yaml with [1]'s and beyond.

[1] https://github.com/theia-ide/trace-server-protocol/blob/master/API.yaml
[2] https://github.com/theia-ide/trace-server-protocol/issues/61
[3] https://github.com/swagger-api/swagger-core/wiki/Swagger-2.X---Annotations#open-api-specification-annotations
[4] https://github.com/theia-ide/trace-server-protocol/pull/62

Change-Id: I9fff048e848ed00b92dae6605effc01c08e3537c
Signed-off-by: Marco Miller <marco.miller@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/186025
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>