server: Shift inlined EndpointConstants w/ imports
DataProviderService was cluttered with all of the EndpointConstants it
had inlined. Replace these with statically imported EndpointConstants,
to shorten those lines and make that code more sustainable.

This change prepares the ground also for the upcoming EndpointConstants,
to be added within many more Swagger annotations.

Organize all imports using Eclipse and this project's own formatting of
them.

Change-Id: I83ca7dfbe0e4f11202c74967b257a8e554c8ae53
Signed-off-by: Marco Miller <marco.miller@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/186557
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>