server: Add Swagger documentation for getProvider
Add documentation to DataProviderService's getProvider endpoint. Base
the added content on [1]'s as usual, but with the necessary adjustments.

Namely, [1]'s getOutput endpoint has been renamed to getProvider since
manually documented that way. Also, the 200 response is no longer made
of Experiment and OutputDescriptor components, but rather a structure
made of the added IDataProvider properties solely. These properties are
a subset of the previously documented OutputDescriptor ones ([1]); they
match with IDataProviderDescriptor's, which is returned by getProvider.

Fix a previous javadoc copy-paste mistake that likely came from the
getProviders sibling method, while working on getProvider this way.

[1] https://github.com/theia-ide/trace-server-protocol/blob/master/API.yaml

Change-Id: I64a7df702d6c527b8510b68f9cccf81968e9bf7e
Signed-off-by: Marco Miller <marco.miller@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/186773
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>