server: Add getTimeGraphTooltip swagger-core doc
Add documentation to DataProviderService's getTimeGraphTooltip endpoint,
based on [1] (below) unless correcting typos or slightly adjusting text.

Among others, introduce IElement to match [1]'s Element entity, which
happens to also match QueryParametersDeserializer's implementation.

In order to match [1] at this level too, introduce ITimeGraphTooltip,
used by ITimeGraphTooltipResponse. The latter returns the former as its
model, similarly with sibling Response interfaces, but replaces the
implemented Map with a List. This is to match [1]'s documented Swagger
array of key-value pairs, which was likely deemed close enough to that
otherwise anonymous Map (of String keys and values). Otherwise, Swagger
interprets the Map as an object made of 'additionalProperties', without
documenting the latter as actual "keys" and "values" (explicitly so).

[1] https://github.com/theia-ide/trace-server-protocol/blob/master/API.yaml

Change-Id: Ib0adfb2f04dd1df5f378851cd4aee6d8848ac7d5
Signed-off-by: Marco Miller <marco.miller@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/188714
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>