Add parser for header of binary FTrace
This commit adds a parser to natively parse the content of the header of
binary FTrace files and its initial tests. The supported FTrace version
is v2.6. Other versions default to trace-cmd reports (legacy way of
working).

Test traces for binary FTrace are stored in the
binary_ftrace_test_trace folder of the test package, due to
conflicts with existing tests. Existing data for FTrace needs to be
restructured first, then all the test traces can be moved into the same
folder.

Currently, trace compass uses trace-cmd to convert binary FTrace traces
into text traces, and then processes it. It requires trace-cmd to be
installed on the running machine. Parsing the binary files directly will
allow trace reading to be platform independent.

Conflicts:
	common/org.eclipse.tracecompass.incubator.target/tracecompass-incubator-master.target

Change-Id: I3d42c9145750cb6d4b3c5a269ea601861d67d5cd
Signed-off-by: Hoang Thuan Pham <hoang.pham@calian.ca>
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/188981
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/190881
Reviewed-by: Marco Miller <marco.miller@ericsson.com>