server: Add swagger-core documentation for /traces
Add documentation to TraceManagerService's endpoints, based on [1] below
unless adjusting text. Following are the summarized noteworthy amends.

Add 'get' to /traces/{uuid}, which maps to getTrace but is not in the
manually documented [1].

Remove [1]'s mention of the uploaded trace URI or path being optional.
Base this removal on the putTrace endpoint method internally expecting a
path passed as parameter. As such, [1] does mention that URI as being
'required', elsewhere in its yaml; keep this herein.

Remove [1]'s redundant http error prefixes from response descriptions.
I.e., do not mention "Not acceptable" or "Conflict" in the latter, since
companion http error codes already convey that information (406, 409).

Remove [1]'s deleteTrace and removeCache parameters, as they are no
longer implemented by the deleteTrace endpoint method. Document that the
trace deletion on disk is included (always, as currently implemented).

Add a 501 response handling even if not in [1]; deemed important enough.

Mention that these response constants are either reused or at least
centralized, in EndpointConstants, as not all of them are reused (yet).

[1] https://github.com/theia-ide/trace-server-protocol/blob/master/API.yaml

Change-Id: Ib6221547e37b6ac1efeebbce9f6f81f62630c767
Signed-off-by: Marco Miller <marco.miller@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/189066
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>