server: Hide QueryParameters from swagger-core doc
Without this change, QueryParameters and its Filter delegate end up in
the openapi.yaml documentation generated by swagger-core. It is so as
QueryParameters remains the type used for actual endpoint method (query)
parameters. This is normal and expected, hence simply hide from Swagger.

This change however (swagger-core) is unable to remove these specific
few openapi.yaml lines:

QueryParameters:
  type: object

Therefore, a companion PR [1] is necessary to remove the above two
lines, as a post-processing step in TSP. An attempt was made to try the
@Hidden annotation at QueryParameters class level, without more success.

[1] https://github.com/theia-ide/trace-server-protocol/pull/72

Change-Id: Ide5f1841e8c340a9bd561cea399436078da35217
Signed-off-by: Marco Miller <marco.miller@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/189088
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>