OTF2: Fix bug in communicators view with asynchronous receive
When there are asynchronous receive operations, 2 events are written in
the OTF2 trace. First there is an IRecvRequest event, then, when the
request is tested in the MPI program, there is the associated IRecv
event containing the information about the communicator. The
communicator view was bugged for these events and would not display the
first event.

To fix it, the information about the request and all the other events
following the unresolved request are stored until the IRecv event is
encountered. Then, the state system is updated with the informations
on the IRecvRequest, the following events and finally the IRecv event.
A trace with an IRecv event can be found here:
https://github.com/dorsal-lab/OTF2_testcases

Signed-off-by: yoann-heitz <yoann.heitz@polymtl.ca>
Change-Id: I9b8b1b5c1c1b3826aae3a5a1d49c01279a257971
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/189385
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Reviewed-by: Marco Miller <marco.miller@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/190888