OTF2: Add a function to modify number of pending locations
Before this patch, to store the changes in the number of pending
locations at given timestamp for each communicator, the put() method
was used on a SortedMap. However, if 2 ranks call a MPI routine within
the same communicator at the same timestamp, this method will be called
two times with the same key, the second time it will erase the
information stored at the first call. A function to increment the
values in this SortedMap is implemented to avoid this behavior when
several changes occur at the same timestamp. The merge() method is used
to increment the value if it already exists or fill the map entry with
the new value if no other value exists.

Signed-off-by: yoann-heitz <yoann.heitz@polymtl.ca>
Change-Id: I9a3eafa986a662fe10ebbe6c933ccf2eddcc797d
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/189386
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>