OTF2: Create generic class representing a SystemTreeNode
Added a generic class representing system tree nodes. This class can be
inherited by the different OTF2 analyzes that use the system
architecture. With this class, some recurrent functions related to
system tree nodes do not need to be duplicated in these analyzes.

Also changed the callstack state provider so it makes use of this new
class.

Signed-off-by: yoann-heitz <yoann.heitz@polymtl.ca>
Change-Id: Ic62bc6ed55dd180adf7f92922d93996446e2c5cc
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/190902
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Marco Miller <marco.miller@ericsson.com>