OTF2: Create generic class representing a LocationGroup
Added a generic class representing location groups (usually MPI
processes). This class can be inherited by the different OTF2 analyzes
that use the system architecture. With this class, some recurrent
functions related to location groups do not need to be replicated.

Also changed the callstack state provider so it uses this new class.

Signed-off-by: yoann-heitz <yoann.heitz@polymtl.ca>
Change-Id: I62a08779de96ccfb6ab68fdcc4e6ce32ae9bfb21
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/190904
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Marco Miller <marco.miller@ericsson.com>