Bug 578841: Fix unknown trace opening with FTrace
No longer pass a negative confidence value from TrimmedFtraceTrace,
while validating if the trace to open is an FTrace or not. This led to
an IllegalArgumentException from TraceValidationStatus that was failing
the opening of a non-FTrace file.

Confidence towards being an FTrace is zero (0) for this specific case,
so this fixed line of code passed a negative confidence value prior. A
complete affected use case description (with steps) is in Bug 578841.

This fix depends on [1]'s own NumberFormatException fix.

[1] https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/190837

Change-Id: I8546cf870ca088f1176685df292b334dec86fe04
Signed-off-by: Marco Miller <marco.miller@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/190968
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
(cherry picked from commit 0429681e75811b24cc65ead3c2f964068dc97c09)
Reviewed-on: https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/190418
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>